Implemented Nodes for 2D Games
As of this commit, PACgd allows to create 2D point and click games.
I had to implement new nodes which are specific for 2D scenes. This
is because 2D elems need to inherit from "Node2D", which has some
subtle yet important differences with "Spatial". First, and most
obviously, Node2D's nodes have positions/sizes as Vector2. But
second, and most importantly, they don't need to deal with cameras
and projections.

The Point and Click main class (pointandclick.gd) remains unchanged,
it is mostly the Interactive class that was modified.

This commit further adds code improvements and fixes to errors that
appeared while testing this new version.