Fixed getting object under mouse
In the last update I introduced a bug when searching for an object
under the mouse. Basically, once an object was decided to be
"avoided", it was going to be avoided forever. I was misleaded by
my own code since the 'pointandclick' class had a property 'avoid'
which I thought was being overwriten from time to time. I was
wrong.

I changed 'avoid' from property to local variable, and now it
works as expected